+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * BUGFIX      #616 Fixed label tick in firefox
     * ENHANCEMENT #615 Introduced no-img-icon as function to use record dependent icons
     * BUGFIX      #613 Fixed search-icon to prevent routing
+    * FEATURE     #610 Added save parameter for datagrid
     * FEATURE     #608 Allows each row in the matrix component to have different values
     * ENHANCEMENT #595 Added default label for is-native selects
     * ENHANCEMENT #594 Changed style of warning label